pyproject.toml
[project]
name = "chanina"
version = "0.2.1"
description = " Chanina is a Python application to automate and orchestrate Playwright sessions using a Celery-based task system. "
authors = [
    {name = "Leo Hehlen",email = "leo.hehlen@pc-rtx.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "playwright (>=1.52.0,<2.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
]

[tool.poetry]
packages = [{include = "chanina", from = "src"}]

[tool.poetry.scripts]
tests = "tests.tests:run"
workflow = "tests.workflow:run"
cli = "src.chanina.cli.runner:run"
docs = "build_docs:run"

[tool.poetry.dependencies]

[tool.poetry.group.docs.dependencies]
sphinx = "^8.2.3"
sphinx-rtd-theme = "^3.0.2"
insegel = "^1.3.1"
alabaster = "^1.0.0"
furo = "^2025.9.25"
sphinx-material = "^0.0.36"
pydata-sphinx-theme = "^0.16.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
